Add files via upload
The JINT4NEOS component allows user to run JavaScript code on an interpreter on the host in it's own thread. Integrating JINT4NEOS into NEOS would allow simple expression to be evaluated quickly as well as implement complex behaviours for NPCs. It is also possible to run client side JavaScript code but this feature has not yet been implemented. 
The threaded program can run continuously in the background and data can be exchanged between the main program and the interpreted program. Data is exchanged between the programs as a series of JSON objects. It also possible to buffer up to 10000 results from the javascript program.
Implementing this feature would allow creators to embed text and data with a NEOS object and to implement complex behaviours such as NPCs that can follow an avatar around a game world,  or process JSON objects for incoming web servers. It also also possible to expose the CLR to JINT, but this feature is turned off for security reasons.