Add files via upload
Added a couple of fields to the InterpretedJS component. You can now set the maximum number of executable instructions  MaxNoOfInstructions field and the maximum. Set MaxNoOfInstructions to make the javascript program run forever ( please use this with caution ).You can also set the amount of usable memory MaxMemoryKB ( although this is implemented as setting it it crashes NEOS ).